# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Inkwell API
  version: "0.2"
  description: |
    API-first Markdown publishing platform.

    All write endpoints require `X-Api-Key`. Reads are public unless stated
    otherwise. See [docs/API.md](./API.md) for the full narrative reference.
  contact:
    name: Inkwell
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your-railway-domain.up.railway.app
    description: Production (replace with actual domain)
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Shared admin key (`INKWELL_API_KEY`) or a scoped token
        (`ink_<prefix>_<secret>`).
  schemas:
    Document:
      type: object
      required:
        - id
        - slug
        - title
        - bodyMarkdown
        - renderedHtml
        - status
        - growth
        - tags
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        slug:
          type: string
          example: "hello-world"
          maxLength: 200
        title:
          type: string
          example: "Hello World"
          maxLength: 500
        bodyMarkdown:
          type: string
          example: "# Hello World\n\nBody text here."
          maxLength: 262144
        renderedHtml:
          type: string
          example: "<h1>Hello World</h1>\n<p>Body text here.</p>"
        status:
          type: string
          enum: [draft, published]
        growth:
          type: string
          enum: [seedling, budding, evergreen]
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
          example: ["rust", "notes"]
        version:
          type: integer
          format: int64
          description: Monotonic revision counter. Echo as `If-Match` on conditional updates.
          example: 3
        createdAt:
          type: string
          format: date-time
          example: "2026-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-06-01T08:00:00Z"
    DocumentList:
      type: object
      required: [documents, total, limit, offset]
      properties:
        documents:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    CreateDocumentRequest:
      type: object
      required: [title, bodyMarkdown]
      properties:
        title:
          type: string
          maxLength: 500
          example: "Hello World"
        bodyMarkdown:
          type: string
          maxLength: 262144
          example: "# Hello World\n\nBody text here."
        slug:
          type: string
          maxLength: 200
          description: Auto-derived from title if omitted.
          example: "hello-world"
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
          example: ["rust", "notes"]
        growth:
          type: string
          enum: [seedling, budding, evergreen]
          default: seedling
...
    AskRequest:
      type: object
      required: [q]
      properties:
        q:
          type: string
          maxLength: 1000
          example: "What is a digital garden?"
    AskResponse:
      type: object
      required: [query, answer, citations]
      properties:
        query:
          type: string
        answer:
          type: string
        citations:
          type: array
          items:
            type: object
            required: [slug, title]
            properties:
              slug:
                type: string
              title:
                type: string
...
    MediaUploadResponse:
      type: object
      required: [id, url]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          example: "/media/550e8400-e29b-41d4-a716-446655440000"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
              example: "Missing or invalid API key."
            requestId:
              type: string
              format: uuid
              description: Correlation ID. Present on every in-request error.
            slug:
              type: string
              description: Present on 409 Conflict for duplicate slug.
            allow:
              type: string
              description: Present on 405 Method Not Allowed.
              example: "GET, POST"
  responses:
    Unauthorized:
      description: Missing or invalid `X-Api-Key`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Missing or invalid API key."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    RejectedReadCredential:
      description: >
        An `X-Api-Key` was presented and rejected (revoked token, unknown prefix,
        hash mismatch, a value that is neither the shared key nor token-shaped, or
        a duplicated/non-ASCII header). Treat this as "re-authenticate", not
        "empty result" — see ADR 0015. A rejected `inkwell_session` cookie is NOT
        reported here: that channel downgrades to anonymous and returns `200` with
        published content, so a stale browser cookie never breaks public pages. A
        valid credential lacking the `read` scope also stays `200`, published only.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Missing or invalid API key."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    ServiceUnavailable:
      description: >
        The credential could not be judged because the token lookup failed against
        Postgres (ADR 0015). Distinct from `401` — the credential was never
        actually rejected — and from `500`. Retry; check `GET /readyz`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Service temporarily unavailable. Retry shortly."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    Forbidden:
      description: Valid credential, but insufficient scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "This action requires the \"write\" scope."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
...
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Rate limit exceeded. Slow down and retry later."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
paths:
...
  /documents:
    get:
      summary: List documents
      operationId: listDocuments
      tags: [Documents]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, published, all]
          description: Filter by status. ANDs with visibility — anonymous callers cannot see drafts.
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: Paginated document list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentList"
        "400":
          description: Invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create a document
      operationId: createDocument
      tags: [Documents]
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
            example:
              title: "Hello World"
              bodyMarkdown: "# Hello World\n\nBody text here."
              tags: ["rust", "notes"]
              growth: "seedling"
      responses:
        "201":
          description: Document created (status is `draft`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Duplicate slug.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  message: "A document with slug \"hello-world\" already exists."
                  slug: "hello-world"
                  requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
        "413":
          description: Request body too large (> 1 MB).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /ask:
    get:
      summary: AI question answering (GET)
      operationId: askGet
      tags: [AI]
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            maxLength: 1000
          example: "What is a digital garden?"
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: Grounded answer with citations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskResponse"
              example:
                query: "What is a digital garden?"
                answer: "A digital garden is an evolving collection of interconnected notes…"
                citations:
                  - slug: "digital-garden"
                    title: "Digital Garden"
        "400":
          description: Empty or oversized query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    post:
      summary: AI question answering (POST)
      operationId: askPost
      tags: [AI]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
            example:
              q: "What is a digital garden?"
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: Grounded answer with citations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskResponse"
        "400":
          description: Empty or oversized query, or invalid JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /media:
    post:
      summary: Upload a media file
      operationId: uploadMedia
      tags: [Media]
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
          image/gif:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Media uploaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResponse"
        "200":
          description: >-
            These exact bytes were already uploaded by this author; the existing
            media is returned. Uploads are content-addressed and idempotent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResponse"
        "400":
          description: >-
            Unsupported Content-Type, or bytes that do not match the declared
            type (the magic bytes are verified against the declaration).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          description: Body exceeds INKWELL_MEDIA_MAX_BYTES (default 5 MiB).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...